LU-1158 general: support for nanosecond m/a/ctimes in Lustre
Extra fields were added in various structures (mostly ones sent over
wire, e.g. mdt_body) for passing around nanosecond m/a/ctimes. Since
ldiskfs/ext4 and the Linux kernel already supported nanosecond
timestamps, it was only a matter of filling in and copying the new
fields in on-wire structures with nanosecond times the rest of the
system provided. Superblocks in the Linux VFS for Lustre also had to
have their s_time_gran fields modified as well to reflect the new time
granularity, and a new OBD connect flag has been added to reflect
support for nanosecond timestamps. Finally, the 39X series of tests in
sanity.sh have been modified to check nanosecond time coherency as
well.

One crucial point to note is that the ost_lvb structure had to be
modified in order to pass along nanosecond times, which required
increasing its size. Since the LVB is apparently not too friendly to
such changes, it is possible (likely?) that this patch will break
compatability with previous versions of Lustre without the addition of
another patch to support variable LVB sizes.

Test-Parameters: fortestonly
Signed-off-by: Isami Romanowski <isami@whamcloud.com>
Change-Id: Ie392d153ae213330dfff6e14bfa2c568959a8ff1