LU-1604 osd: don't call md_ucred() for OFD
There are several places where md_ucred() is called in osd-ldiskfs.
However, this function is specific to the metadata stack (it ends
up calling a callback in mdd layer) and should not be called with OFD.

osd_{push,pop}_ctxt() have thus been modified to be no-op when the top
device is OFD. __osd_oi_insert() has also been changed to call
md_ucred() only when dealing with a MDT device.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: I931f0376a78cb2aff8b00a99b1c20618d4093e7d