LU-1608 utils: require that fsname is specified
When formatting a filesystem, require that fsname is specified to
mkfs.lustre so that users have to think about this a bit.  In some
cases, two filesystems at the same site have used the same default
filesystem name, and this later caused problems when trying to mount
both filesystems on the same node.

Improve error messages in ldd_parse(), ldd_write(), and ldd_print()
to include the device/target name, so that if these ever print on a
multi-target server it will be possible to see which one failed.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I15a5ebb8d7d3afc6037587961e94dc66426bcf5d