LU-1562 test: recovery-small tests 2,3,5 failed
1. The code that executed between tests has been cleaned up
and moved to corresponding tests which use it to avoid its
execution all the time, even if it is not required when
running using ONLY.
2. The test_2, and test_3 has been joined with test_1,
and the test_7 has been joined with test_6, because
they are in fact single scenario.
3. Random and zero file creation code moved to corresponding
utility functions in the test-framework.sh
4. Added bash stack trace printing code to error_noexit that
make it easier to analyze logs.

Xyratex-bug-id: MRP-411
Signed-off-by: Kyrylo Shatskyy <kyrylo_shatskyy@xyratex.com>
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>
Reviewed-by: Alexander Zarochentsev <alexander_zarochentsev@xyratex.com>
Change-Id: Iea417c99edb001e4f5ad0d0980c7c10012e2514a