LU-1622 lnet: hash MEs on wildcard portal
If multiple services share one portal but use match-bits to identify
different buffer types, we still have performance issue because all
buffers are attached on one list and searching for typed buffer
could be long and expensive. So we should hash buffers on wildcard
portal as well.

Another issue addressed by this patch is, we should always attach
MEs with ignore-bits on a list instead of hash-table, no matter it's
wildcard portal or unique portal, because message could match buffers
with various match-bits if they also had ignore-bits.

Signed-off-by: Liang Zhen <liang@whamcloud.com>
Change-Id: I0d2c6330dd231d369e2a86ced2b8374c0c96dbf9