LU-1497 client: SLES11 SP0 client build is broken.
With the addition of client improvements new linux kernel
apis are exposed to the llite layer. Those apis can be
missing or have a different format in earlier kernels.
This is the case for SLES11 SP0 clients for the case
of the vma handling and the function inode_newsize_ok
that is missing. This patch fixes those issues. Final
bug is a fix for test a function for the crypto api.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I699c67baec030e61133ff0a44e0ab03261b2f2d0