LU-1606 api: Rename liblustreapi.h -> lustreapi.h
The header file "liblustreapi.h" is unfortunately named,
implying that it is only part of the lustre-in-user-space-client
rather than the more general lustre api for users that it is.

This patch renames it to simply "lustreapi.h" to avoid that
confusion.  This also helps to move towards making it
easier to use the lustre API for users.  They can simply

  #include <lustre/lustreapi.h>

and then compile with "-llustreapi".

For backwards compatibility, liblustreapi.h becomes a stub
that includes lustreapi.h, and can be removed at some date
in the future.

Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Change-Id: Id1e709d36a855ad2a8eff206fa9f6bbe87182a29