LU-1606 api: Move defines from lustre_idl.h to lustre_user.h
Move some defines that user's require to be able to use
the api's get_stripe functions.  Since lustre_user.h is the
place designated in lustre_idl.h as the place for definitions
needed both in the IDL and in user space, and since lustre_idl.h
will not compile in user space, it seems logical to relocate
these definitions there.

Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Change-Id: Ib0133553f902d68222c16c5fc5b99cd5a9cd033d