LU-1623 mdc: Ignore unexpected server remote acl data
Somethign strange is going on on the MDS that causes it to think
clients are "remote" and sending them some relevant ACL data.
Now, the clients are really surprised by such turn of events and as such
trigger an assertion.

The root cause is unknown and the crashes annoy people so this is
a temporary workaround patch to stop such assertion from triggering.

Change-Id: I5b9e6c72fda403c51076af1bee1766851f3846b4
Signed-off-by: Oleg Drokin <green@whamcloud.com>