LU-1629 ldlm: Fix recovery messages when denying new clients
Console messages printed when denying new client connections during
target recovery show misleading client statistics.  For example, in
this case:

  Lustre: lustre-MDT0000: Denying connection for new client
  192.168.117.50@o2ib1 (at 11e711ab-a329-f07a-8312-6a40af7fc5a4),
  waiting for 0 clients in recovery for 2:38
  [...]
  Lustre: lustre-MDT0000: Recovery over after 5:00, of 112 clients 0
  recovered and 112 were evicted.

The MDT was actually waiting for all the 112 known clients to recover.
None had connected, however.  In addition, the client NID and UUID
seem to be in the wrong order.  This patch changes the first console
message to look like this:

  Lustre: lustre-MDT0000: Denying connection for new client
  939243e4-2f54-a96f-3cbb-9fcf55426e2e (at 0@lo), waiting for all 2
  known clients (1 recovered, 0 in progress, and 1 unseen) to
  recover in 0:05

Hopefully this new format will be a little bit more useful to users,
although the counters are (still) read in a racy way.

Change-Id: Iefda085602de7967d66892b8f3567561962078ab
Signed-off-by: Li Wei <liwei@whamcloud.com>