LU-1581 mount: use a special separator in svname
with this new approach mkfs marks fresh filesystem
by a special separator (:) between fsname and service
name. then mount utility recognizes this separator
and passes 'writeconf' option to the kernel mount.
upon this first mount (and having a new configuration
profile on MGS) mount utility replaces ':' with regular
'-' in svname and sets label to that.

Signed-off-by: Alex Zhuravlev <bzzz@whamcloud.com>
Change-Id: I83e1d1e4f1fd67ae9a6453069b324696715d4fee