LU-445 lnet: Send timestamps with LNet counters b2_2
We use a millisecond timer on a per-session basis. Each node then
sends the milliseconds since session start over with the LNet
counters. With 32bits and millisecond resolution, this should
allow ~49 days of session age before the timers would wrap.

This will change the wire attribute and not backward compatible.
In order to maintain backward compatibility, this patch checks
the value of the timestamp field to determine if it is a
valid value (>= 2.2) or is too small to be a timestamp (< 2.2).
If it is a valid timestamp, it will be used. If not, the local
timestamp will be used.

Change-Id: I0dc6498d187ccc900b86d21b0460e1d419770033
Signed-off-by: Wally Wang <wang@cray.com>
Signed-off-by: Liang Zhen <liang@whamcloud.com>