LU-1735 ptlrpc: only set jobid if not already set
The ptlrpc_set_add_req->lustre_get_jobid->cfs_access_process_vm()
callpath locks mm->mmap_sem to fetch environment variables, but if
this could deadlock in case of MMAP IO, which also holds mmap_sem.

The caller already has fetched the jobid and saved it for this inode
in the OSC layer, so we don't need to fetch and reset the pd_jobid
field at all in this case.  This not only avoids the deadlock, but
also avoids doing extra work to fetch the jobid if it is not needed,
and avoids storing it temporarily on the stack when it won't be used.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I22c3d5c1755c1d6aab666a769df38218b954650a