LU-1434 utils: Update Wireshark support for LNet and Lustre
There was work previously done to create two plugins for Wireshark to
decode LNet and Lustre headers. That work is started in lustre/contrib
as the files packet-lnet.c and packet-lustre.c.

Attempting to build these plugins revealed that Wireshark's API for
dissectors has changed such they no longer build/run. This change
fixes the plugins so they work with the latest stable Wireshark
release: 1.6.8. It also adds logic to allow the plugins to decode
Infiniband packets captured with ibdump.

Signed-off-by: Doug Oucharek <doug@whamcloud.com>
Change-Id: I5d7df739eb5d3eec8556856c948a422df72bbfd6