LU-1787 osc: osc_setup doesn't return rc in all cases
The function osc_setup always exits with a rc value. If
ptlrpcd_alloc_work fails in osc_setup the variable rc is
not set with the error code but with zero which reports
sucess which is not the case. The fix is set rc to the
error value return from ptlrpcd_alloc_work if it fails.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I3e3792a52defb5f8c4263ea62b2ea282723e458f