LU-1789 protocol: reserve connect flag for lightweight conn
Reserve connection flag for lightweight connection support to avoid
conflicts. The main property of a lightweight connection is that
there is no entry in the last_rcvd file for this export, so no
recovery is possible. This type of connection will be used by quota
(between master & slave), FIDonOST and likely by other features in the
future.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: I4367f5bfaef0207e7f307c4ea3d35e97ca99dc5b