LU-718 mdt: allow mds_num_threads module parameter
In 1.8.x the MDS service threads tunable is called "mds_num_threads",
but in 2.0/2.1 this was renamed mdt_num_threads without any chance
of compatibility.  In 2.3.x this was named back to mds_num_threads
(commit bd8835cc2dde6e86701650ccf90423ecd8fb042e) since the threads
are a property of the MDS, not of the MDT(s) on the system.

Add a compatible module option "mds_num_threads" for 2.1.x so that
it is possible to upgrade 1.8->2.1->2.4 without problems.  No message
is printed in 2.1.x about deprecation, since it isn't possible to use
only one or the other in 2.1, and a warning will be printed in 2.3+
once the system is upgraded.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I3cb095812488b9459e4a3e878757d40410accab0