LU-1430 mgs: "replace_nids" lctl command added
Currently after device address is changed tunefs
should be started with "writeconf" options. This
erase configuration logs for the filesystem that
this server is part of, and regenerate them. This
is very dangerous. All clients must be unmounted
and servers  for  this filesystem should be stopped.
All targets (OSTs/MDTs) must then be restarted to
regenerate the logs.  No clients should be started
until all targets have restarted. The targets start
order is also important. Wrong order can be cause
of failure.

This patch adds "replace_nids" command to lctl
that should be executed when only MGS server
started. This command replace nids and uuid for
given device. Command can be used to change
OST or MDT nids.

Xyratex-bug-id: MRP-397
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Reviewed-by: Nathan Rutman <nathan.rutman@xyratex.com>
Signed-off-by: Artem Blagodarenko <artem_blagodarenko@xyratex.com>
Change-Id: I74b72281ce3b5e48188735035e58250596a16bf7