LU-1935 scripts: don't generate lustre init script
An "overriding commands for target `lustre'" build warning resulted
from the addition of lustre to the genscripts Makefile variable.  In
fact the genscripts processing is superfluous for the lustre script,
so we could just fix the warning by removing it.  However, generating
the script from a .in file seems unnecessarily complicated, so this
change goes a step further by making the script static.  It now
detects at run-time whether to use the ldiskfsprogs or the e2fsprogs
utilities, instead of using autoconf macros to encode it at build
time.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: Ie000a87adabbc2a001ed9ff2f5750041a5abea5b