LU-1996 lustre: Add jobid to changelogs.
Added jobid fields to Changelog records (and extended records).
The HAS_JOBID flags allows to check if the field is present or
not (old format) when reading an entry.

Jobids are expressed as 32 chars long, zero-terminated strings.

Introduced a related test case to sanity.sh.
Adjusted test_160 to handle new Changelog format.

Signed-off-by: Henri Doreau <henri.doreau@cea.fr>
Change-Id: Iacfe7099aa520e012535ac48c44793a1b86870a9