LU-2068 build: Respect "--without-ldiskfs" option
The "--without-ldiskfs" configure option was not properly being
respected by a few of the Makefiles. This change is an attempt
to remedy this issue. Now, one should be able to disable building
ldiskfs and the related lvfs and osd code with "--without-ldiskfs".

Also, with the option obdfilter, llog_lvfs are not built, mkfs
utility does not call to ldiskfs helpers.

Signed-off-by: James Simmons<uja.ornl@gmail.com>
Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I502e125639a60f62a37e3f95ecc943dedc37f5ea