LU-1175 tests: use ofd by default instead of obdfilter
Set USE_OFD=yes by default in order to start testing the ofd module in
preference to the obdfilter module.  Force LOAD_MODULES_REMOTE to be
set to avoid obdfilter being loaded by modprobe remotely on mount.

There was a bug in comma_list() appending a space to the list of
nodes, which caused do_rpc_nodes() to be unhappy when a nearly-empty
list of nodes was being passed by callers.  Rework this function, and
clean up other callers to just pass the nodes and let do_rpc_nodes()
detect this case and skip any remote execution.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I896c9c9b7216c418b635e893202810333d8ef871