LU-790 tests: clean up lustre version checking
Since commit 48bad5d9db9baa7bca093de5c54294adf1cf8303, the check in
sanity-scrub.sh to verify the MDS had a new enough Lustre version was
being run before modules are loaded, if sanity-scrub.sh was being run
stand-alone.  Instead of using "lctl get_param version" to get the
Lustre version, use "lctl lustre_build_version", which will always
work, even if the modules are not loaded on the node.  This also
avoids the need to load the modules just to determine that the test
should not be run.

While cleaning up this code, move the lustre_version_code() earlier in
test-framework.sh, so that it is declared before calling functions,
and is co-located with the other version checking code.  Not strictly
necessary for bash scripts, but nicer when reading the code.

Also remove the obsolete check_kernel_version() code, which was only
used when we had a patched client kernel, which is no longer the case,
and remove a few overly verbose logging lines from tests.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I8a8aec11370d395dcb4f4b250001ecdb40500c1e