LU-1095 debug: Log expired requests as D_NETERROR
While knowing about expired requests is important, it's also not
an uncommon event on a large system.  As long as the lower layers
can retry or take other corrective action there's no reason to
both an administrator with this.  They can always set D_NETERROR
is they suspect problems with the network.

Signed-off-by: Minh Diep <mdiep@whamcloud.com>
Change-Id: I07044a31aa28aca9a7b965b6cd9e237e485d5548