LU-1538 tests: use $TESTSUITE instead of $0
Use "$TESTSUITE" instead of "$0" in test script messages, since $0 is
a full pathname and clutters up the test logs.  Instead, $TESTSUITE is
only the test suite name, and is more compact.  Use it in all output.

This makes the output from complete() redundant, in that it prints
$TESTSUITE both as an argument and internally via the equals_msg()
function (via banner()), so remove the $TESTSUITE argument from all
callers of complete().

The equals_msg() function was only a thin wrapper around banner(), and
was only used by complete(), remove it and call banner() directly.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I6b3f256000317a17fdd2a361a38d4dfdda500c1e