LU-1538 tests: standardize test script initialization
Standardize the initial Lustre test script initialiazation of the
test-framework.sh for clarity and consistency.

The LUSTRE path is already normalized in init_test_env(), so this
doesn't need to be done in the caller.  Use $(...) subshells instead
of `...` in the affected lines.  Remove PATH, NAME, TMP, LFS, LCTL
variable initialization, since it is already done in init_test_env().

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I190408ffdb4c19a7be4635713cedf28f5d500c1e