LU-1623 Atomically update server export connection flags
It appears that we update export connection flags on MDS in two step.
Step one is simple feature matching and then much later we also
remove various security-related bits as needed.
In the case of reconnect some requestp processing can see interim
state of the flags and assuem client has capabilities that it does
not really have and send some confusing replies to it.

This patch moves all exp_vonnection_flags assignment in mdt to
mdt_init_sec_level only.

Change-Id: I3a6b16a589083feeb4ab166625648cd7a1652793
Signed-off-by: Oleg Drokin <green@whamcloud.com>