LU-2139 osc: Limit number of "unstable" on client
This change adds a per filesystem limit on the number of "unstable"
pages held by the client. An "unstable" page is defined as page which is
sent to the server as part of a bulk request, but is uncommitted to
stable storage.

The current number of, and maximum allowed, pages are stored in the
ll_sb_info structure. When the current number of unstable pages reaches
the maximum, calls to osc_page_init will sleep until pages are committed
on the server. Thus reducing the number of unstable pages on the client.

The motivation for this change is to prevent clients from allocating too
much of its available RAM to unreclaimable pages pinned by the ptlrpc
layer. If this case occurs, out of memory events can trigger as a side
effect, which we need to avoid.

The number of unstable pages tracked in the ll_sb_info structure are
exported by the max_unstable_mb proc file, contained under each
filesystem's llite namespace. This file can also be written to, allowing
the upper bound of unstable pages on the system to be modified
dynamically.

For example:

    # Set unstable page limit to 512 MB
    echo 512 > /proc/fs/lustre/llite/lustre-ffff88004e338400/max_unstable_mb

    # Read unstable limit, number of pages, and number of waiters
    cat /proc/fs/lustre/llite/lustre-ffff88004e338400/max_unstable_mb

In addition to tracking the number of pages in the ll_sb_info structure,
the NR_UNSTABLE_NFS memory zone is also incremented and decremented
for easy monitoring using the "NFS_Unstable:" field in /proc/meminfo.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Ic43d34bfa39da9ac4fe159000c7db5908467fd7b